requirements.txt
fastapi
uvicorn[standard]
pydantic
firebase-admin
python-dotenv
cryptography
torch
transformers
google-cloud-firestore
